+++ example/lib/gen/voyager_gen.dart
@@ -1,6 +1,5 @@
 /// Generated file, DO NOT EDIT
 import 'package:flutter/widgets.dart';
-import 'package:provider/provider.dart';
 import 'package:voyager/voyager.dart';
 
 const String pathHome = "/home";
@@ -20,10 +19,7 @@ String pathObjectItem(String className) {
 
 const String typeObjectItem = "object_item";
 
-class VoyagerData extends Voyager {
-  VoyagerData({String path, Voyager parent, Map<String, dynamic> config})
-      : super(path: path, parent: parent, config: config);
-
+extension VoyagerData on Voyager {
   String get title => this["title"];
   String get body => this["body"];
   String get fabPath => this["fabPath"];
@@ -32,17 +28,6 @@ class VoyagerData extends Voyager {
   List<dynamic> get actions => this["actions"];
   List<dynamic> get items => this["items"];
 }
-
-class VoyagerProvider {
-  static VoyagerData of(BuildContext context) => Provider.of<Voyager>(context);
-}
-
-VoyagerData voyagerDataFactory(
-        AbstractRouteContext abstractContext, Map<String, dynamic> config) =>
-    VoyagerData(
-        path: abstractContext.url(),
-        parent: abstractContext.getExtras().parent,
-        config: config);
 
 abstract class IconPluginStub extends RouterObjectPlugin<Icon> {
   IconPluginStub() : super("icon");