+++ example/test/widget_test.dart
@@ -1,4 +1,3 @@
-import 'package:example/gen/voyager_gen.dart';
 import 'package:example/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -75,7 +74,6 @@ class TestScenarios extends VoyagerTestScenarios {
 }
 
 void main() {
-  final router =
-      loadRouter(paths(), plugins(), voyagerFactory: voyagerDataFactory);
+  final router = loadRouter(paths(), plugins());
   voyagerAutomatedTests("voyager auto tests", router, TestScenarios());
 }